add role guard test to form
