+++ test/utils.ts
@@ -1,3 +1,9 @@
+import { PrismaClient } from "@prisma/client";
+import { INestApplication } from "@nestjs/common";
+import * as request from "supertest";
+
+const prisma = new PrismaClient();
+
 // returns just the token
 export const extractResCookieValueByKey = (cookies, key) => {
     return cookies
@@ -10,4 +16,43 @@ export const extractResCookieValueByKey = (cookies, key) => {
 // returns 'access_token={token}; Max-Age=1800; Path=/; Expires=Fri, 16 Feb 2024 03:17:25 GMT; HttpOnly; Secure'
 export const extractCookieByKey = (cookie, key) => {
     return cookie.filter((cookie) => cookie.includes(key))[0];
+};
+
+export const loginAndGetTokens = async (
+    email: string,
+    password: string,
+    app: INestApplication,
+) => {
+    const r = await request(app.getHttpServer()).post("/auth/login").send({
+        email,
+        password,
+    });
+
+    const access_token = extractCookieByKey(
+        r.headers["set-cookie"],
+        "access_token",
+    );
+    const refresh_token = extractCookieByKey(
+        r.headers["set-cookie"],
+        "refresh_token",
+    );
+
+    return { access_token, refresh_token };
+};
+
+export const getNonAdminUser = async () => {
+    const adminRole = await prisma.role.findUnique({
+        where: {
+            name: "admin",
+        },
+    });
+    return prisma.user.findFirst({
+        where: {
+            roles: {
+                none: {
+                    roleId: adminRole.id,
+                },
+            },
+        },
+    });
 };
